First attempt at prio.act
Linear buffer, one input for data, one input for command, one output
Used an approach where the insert function just sticks the new number into the back of the linear buffer (passes into the input). 
The remove function will cycle through all the numbers in the buffer and determine the lowest number, then output. Thus, we're never actually ordering the numbers in the buffer; we're just finding the smallest number in the buffer every time we want to output something.